ORA-29 Frontend request page fix + minor improvements
